fetcher/pyproject.toml
[tool.poetry]
name = "fetcher"
version = "0.1.0"
description = ""
authors = ["Russ Garrett <russ@garrett.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
databases = "^0.6.1"
aiopg = "^1.3.5"
httpx = "^0.23.0"
python-dateutil = "^2.8.2"
asyncpg = "^0.26.0"
asyncstdlib = "^3.10.5"
aiostomp = "^1.6.2"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
black = "^22.10.0"
types-python-dateutil = "^2.8.19"
pytest = "^7.1.3"
types-pytz = "^2022.4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
